wip: inductive alphabet types
Isomorphic inductive alphabet types for encodings, for example sigNat instead of simply bool, or sigSum sigX sigY instead of bool+(sigX+sigY).

The intention is to make finding retracts between tape alphabets easier.
For example, if a tape t : bool+(bool + bool) is intended to store a sum of a natural number (previously encoded over bool) and a bool, if we write t \simeq n, it isn't clear to which of the bools the encoding should be mapped.
But if the alphabet is sigSum sigNat bool, then it is clear that the retract should go "into" the sigNat to the left side.