Make the tactic destruct_shelve more general
This is needed so that it can work for alphabet with custom types.
Another aproach could be to use smpl to let the user add a
destruct_shelve tactic for each custom type.

This implementation works for inductive types with up to 9 constructors,
but it can generate a lot of warnings, which can be ignored.