Translate-Machine
This machine translates a value from one possible encoding to another
encoding on the same tape.

For example, if X is encodable over sigma and there are two retracts f1,
f2 from sigma to tau, then there are two ways to encode X on tau: via
Encode_map f1 and via Encode_map f2.

This machine assumes that the value is encoded via the first encoding
and translates it into the second encoding.