TM_Correct instances for Matches and WriteValue
There is no need to apply the correctness lemmas for matches manually.
For example, if we have MatchList sigX, then we assume that there is
no other instance of codable sigX X. (I.e. if codable sigX Y, then X=Y).
This assumption is true in general, so we can also use the TM_Correct
mechanism for matches.

For WriteValue, we have to give the codable instance directly to the
machine, so we can also automatically apply the correctness lemmas of
WriteValue.